requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==4.2
django-environ==0.12.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
phonenumbers==8.13.53
pillow==11.1.0
psycopg2==2.9.10
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.8.2
